+++ photutils/detection/core.py
@@ -39,8 +39,9 @@ class StarFinderBase(metaclass=abc.ABCMeta):
         convolved_data : 2D array_like
             The convolved 2D array.
 
-        kernel : `_StarFinderKernel`
-            The convolution kernel.
+        kernel : `_StarFinderKernel` or 2D `~numpy.ndarray`
+            The convolution kernel. ``StarFinder`` inputs the kernel
+            as a 2D array.
 
         threshold : float
             The absolute image value above which to select sources. This
@@ -83,46 +84,29 @@ class StarFinderBase(metaclass=abc.ABCMeta):
             footprint = np.array((xx**2 + yy**2) <= min_separation**2,
                                  dtype=int)
 
-        # pad the convolved data and mask by half the kernel size (or
-        # x/y radius) to allow for detections near the edges
-        if isinstance(kernel, np.ndarray):
-            ypad = (kernel.shape[0] - 1) // 2
-            xpad = (kernel.shape[1] - 1) // 2
+        # define the border exclusion region
+        if exclude_border:
+            if isinstance(kernel, np.ndarray):
+                yborder = (kernel.shape[0] - 1) // 2
+                xborder = (kernel.shape[1] - 1) // 2
+            else:
+                yborder = kernel.yradius
+                xborder = kernel.xradius
+            border_width = max(xborder, yborder)
         else:
-            ypad = kernel.yradius
-            xpad = kernel.xradius
-
-        if not exclude_border:
-            pad = ((ypad, ypad), (xpad, xpad))
-            pad_mode = 'constant'
-            convolved_data = np.pad(convolved_data, pad, mode=pad_mode,
-                                    constant_values=0.0)
-            if mask is not None:
-                mask = np.pad(mask, pad, mode=pad_mode, constant_values=False)
+            border_width = 0
 
         # find local peaks in the convolved data
         # suppress any NoDetectionsWarning from find_peaks
         with warnings.catch_warnings():
             warnings.filterwarnings('ignore', category=NoDetectionsWarning)
             tbl = find_peaks(convolved_data, threshold, footprint=footprint,
-                             mask=mask)
+                             mask=mask, border_width=border_width)
 
         if tbl is None:
             return None
 
-        if exclude_border:
-            xmax = convolved_data.shape[1] - xpad
-            ymax = convolved_data.shape[0] - ypad
-            mask = ((tbl['x_peak'] > xpad) & (tbl['y_peak'] > ypad)
-                    & (tbl['x_peak'] < xmax) & (tbl['y_peak'] < ymax))
-            tbl = tbl[mask]
-
-        xpos, ypos = tbl['x_peak'], tbl['y_peak']
-        if not exclude_border:
-            xpos -= xpad
-            ypos -= ypad
-
-        return np.transpose((xpos, ypos))
+        return np.transpose((tbl['x_peak'], tbl['y_peak']))
 
     @abc.abstractmethod
     def find_stars(self, data, mask=None):